updated all enums with casts from i32 to u32
bindgen seems to have changed the way, how enums are generated, or it is
platform dependend
on windows bindgen generates i32 but imgui-sys rust code uses u32

if bindgen generates u32 on unix, then a lot of warnings for
unecessary cast might be emitted

And finally a new example "imgui_sys" is added to show, how to use the
evil unsafe igFunctions to create awesome widgets!

+++ imgui/src/fonts/atlas.rs
@@ -17,11 +17,11 @@ bitflags! {
     #[repr(transparent)]
     pub struct FontAtlasFlags: u32 {
         /// Don't round the height to next power of two
-        const NO_POWER_OF_TWO_HEIGHT = sys::ImFontAtlasFlags_NoPowerOfTwoHeight;
+        const NO_POWER_OF_TWO_HEIGHT = sys::ImFontAtlasFlags_NoPowerOfTwoHeight as u32;
         /// Don't build software mouse cursors into the atlas
-        const NO_MOUSE_CURSORS = sys::ImFontAtlasFlags_NoMouseCursors;
+        const NO_MOUSE_CURSORS = sys::ImFontAtlasFlags_NoMouseCursors as u32;
         /// Don't build thick line textures into the atlas
-        const NO_BAKED_LINES = sys::ImFontAtlasFlags_NoBakedLines;
+        const NO_BAKED_LINES = sys::ImFontAtlasFlags_NoBakedLines as u32;
     }
 }
