+++ imgui/src/widget/color_editors.rs
@@ -91,60 +91,60 @@ bitflags! {
     pub struct ColorEditFlags: u32 {
         /// ColorEdit, ColorPicker, ColorButton: ignore Alpha component (read only 3 components of
         /// the value).
-        const NO_ALPHA = sys::ImGuiColorEditFlags_NoAlpha;
+        const NO_ALPHA = sys::ImGuiColorEditFlags_NoAlpha as u32;
         /// ColorEdit: disable picker when clicking on colored square.
-        const NO_PICKER = sys::ImGuiColorEditFlags_NoPicker;
+        const NO_PICKER = sys::ImGuiColorEditFlags_NoPicker as u32;
         /// ColorEdit: disable toggling options menu when right-clicking on inputs/small preview.
-        const NO_OPTIONS = sys::ImGuiColorEditFlags_NoOptions;
+        const NO_OPTIONS = sys::ImGuiColorEditFlags_NoOptions as u32;
         /// ColorEdit, ColorPicker: disable colored square preview next to the inputs. (e.g. to
         /// show only the inputs)
-        const NO_SMALL_PREVIEW = sys::ImGuiColorEditFlags_NoSmallPreview;
+        const NO_SMALL_PREVIEW = sys::ImGuiColorEditFlags_NoSmallPreview as u32;
         /// ColorEdit, ColorPicker: disable inputs sliders/text widgets (e.g. to show only the
         /// small preview colored square).
-        const NO_INPUTS = sys::ImGuiColorEditFlags_NoInputs;
+        const NO_INPUTS = sys::ImGuiColorEditFlags_NoInputs as u32;
         /// ColorEdit, ColorPicker, ColorButton: disable tooltip when hovering the preview.
-        const NO_TOOLTIP = sys::ImGuiColorEditFlags_NoTooltip;
+        const NO_TOOLTIP = sys::ImGuiColorEditFlags_NoTooltip as u32;
         /// ColorEdit, ColorPicker: disable display of inline text label (the label is still
         /// forwarded to the tooltip and picker).
-        const NO_LABEL = sys::ImGuiColorEditFlags_NoLabel;
+        const NO_LABEL = sys::ImGuiColorEditFlags_NoLabel as u32;
         /// ColorPicker: disable bigger color preview on right side of the picker, use small
         /// colored square preview instead.
-        const NO_SIDE_PREVIEW = sys::ImGuiColorEditFlags_NoSidePreview;
+        const NO_SIDE_PREVIEW = sys::ImGuiColorEditFlags_NoSidePreview as u32;
         /// ColorEdit: disable drag and drop target. ColorButton: disable drag and drop source.
-        const NO_DRAG_DROP = sys::ImGuiColorEditFlags_NoDragDrop;
+        const NO_DRAG_DROP = sys::ImGuiColorEditFlags_NoDragDrop as u32;
         /// ColorButton: disable border (which is enforced by default).
-        const NO_BORDER = sys::ImGuiColorEditFlags_NoBorder;
+        const NO_BORDER = sys::ImGuiColorEditFlags_NoBorder as u32;
 
         /// ColorEdit, ColorPicker: show vertical alpha bar/gradient in picker.
-        const ALPHA_BAR = sys::ImGuiColorEditFlags_AlphaBar;
+        const ALPHA_BAR = sys::ImGuiColorEditFlags_AlphaBar as u32;
         /// ColorEdit, ColorPicker, ColorButton: display preview as a transparent color over a
         /// checkerboard, instead of opaque.
-        const ALPHA_PREVIEW = sys::ImGuiColorEditFlags_AlphaPreview;
+        const ALPHA_PREVIEW = sys::ImGuiColorEditFlags_AlphaPreview as u32;
         /// ColorEdit, ColorPicker, ColorButton: display half opaque / half checkerboard, instead
         /// of opaque.
-        const ALPHA_PREVIEW_HALF = sys::ImGuiColorEditFlags_AlphaPreviewHalf;
+        const ALPHA_PREVIEW_HALF = sys::ImGuiColorEditFlags_AlphaPreviewHalf as u32;
         /// (WIP) ColorEdit: Currently onlys disable 0.0f..1.0f limits in RGBA editing (note: you
         /// probably want to use `ColorEditFlags::FLOAT` as well).
-        const HDR = sys::ImGuiColorEditFlags_HDR;
+        const HDR = sys::ImGuiColorEditFlags_HDR as u32;
         /// ColorEdit: display only as RGB. ColorPicker: Enable RGB display.
-        const DISPLAY_RGB = sys::ImGuiColorEditFlags_DisplayRGB;
+        const DISPLAY_RGB = sys::ImGuiColorEditFlags_DisplayRGB as u32;
         /// ColorEdit: display only as HSV. ColorPicker: Enable HSV display.
-        const DISPLAY_HSV = sys::ImGuiColorEditFlags_DisplayHSV;
+        const DISPLAY_HSV = sys::ImGuiColorEditFlags_DisplayHSV as u32;
         /// ColorEdit: display only as HEX. ColorPicker: Enable Hex display.
-        const DISPLAY_HEX = sys::ImGuiColorEditFlags_DisplayHex;
+        const DISPLAY_HEX = sys::ImGuiColorEditFlags_DisplayHex as u32;
         /// ColorEdit, ColorPicker, ColorButton: _display_ values formatted as 0..255.
-        const UINT8 = sys::ImGuiColorEditFlags_Uint8;
+        const UINT8 = sys::ImGuiColorEditFlags_Uint8 as u32;
         /// ColorEdit, ColorPicker, ColorButton: _display_ values formatted as 0.0f..1.0f floats
         /// instead of 0..255 integers. No round-trip of value via integers.
-        const FLOAT = sys::ImGuiColorEditFlags_Float;
+        const FLOAT = sys::ImGuiColorEditFlags_Float as u32;
         /// ColorPicker: bar for Hue, rectangle for Sat/Value.
-        const PICKER_HUE_BAR = sys::ImGuiColorEditFlags_PickerHueBar;
+        const PICKER_HUE_BAR = sys::ImGuiColorEditFlags_PickerHueBar as u32;
         /// ColorPicker: wheel for Hue, triangle for Sat/Value.
-        const PICKER_HUE_WHEEL = sys::ImGuiColorEditFlags_PickerHueWheel;
+        const PICKER_HUE_WHEEL = sys::ImGuiColorEditFlags_PickerHueWheel as u32;
         /// ColorEdit, ColorPicker: input and output data in RGB format.
-        const INPUT_RGB = sys::ImGuiColorEditFlags_InputRGB;
+        const INPUT_RGB = sys::ImGuiColorEditFlags_InputRGB as u32;
         /// ColorEdit, ColorPicker: input and output data in HSV format.
-        const INPUT_HSV = sys::ImGuiColorEditFlags_InputHSV;
+        const INPUT_HSV = sys::ImGuiColorEditFlags_InputHSV as u32;
     }
 }
